Combines pwv_atm module into a single document
